# Match arity of `Hash#default` in `HWIA#default`